Move query helper to winrt_helpers.dart
