+++ lib/src/winrt/winrt_helpers.dart
@@ -174,6 +174,27 @@ enum TrustLevel {
 }
 
 extension IInspectableExtension on IInspectable {
+  /// Queries `IInspectable` with the given `iid` for the relevant interface.
+  ///
+  /// Throws a `WindowsException` if the interface is not found.
+  ///
+  /// It is the caller's responsibility to deallocate the returned pointer when
+  /// they are finished with it. A FFI `Arena` may be passed as a
+  /// custom allocator for ease of memory management.
+  ///
+  /// {@category winrt}
+  Pointer<COMObject> query(String iid, {Allocator allocator = calloc}) {
+    final classPtr = allocator<Pointer>();
+    // Create an IID for the interface required
+    final riid = convertToIID(iid, allocator: allocator);
+    // Use IInspectable to navigate to the relevant interface
+    final hr = this.QueryInterface(riid, classPtr);
+    if (FAILED(hr)) throw WindowsException(hr);
+
+    // Return a pointer to the relevant class
+    return classPtr.cast();
+  }
+
   // TODO: Add iids property when IVectorView`1 work lands.
   List<String> get iids {
     throw UnimplementedError();
